Add basic materials + better config checking
Basic and naive implementation. There is still a lot in the material
file format I don't yet understand, which will require further reverse
engineering.

There seems to be some sort of inheritance in materials, as, for example
many of the enemy materials reference a single common material, which dictates
the exact textures they use. We can definitely use this by finding some
of the common parent materials and defining a way to handle material
conversion to glTF for each of them.

It seems these parent materials contain the actual DirectX shader code,
so we'll essentially have to emulate that code as best as possible in
the glTF format.