Add bones/skins to exported gltf models
Still needs the rotations and translations and bone indices to be fixed up.
Bone indices appear to be offset from the actual values, maybe there's something in the vertex groups that indicates this?
Rotations and translations need to be adjusted to gltf relative coordinates.
Finally, <file>.bones need to be used to map the bone hashes to names