Initial steps towards workflow importing directly into blender with accurate shaders
Blender only has a python API, so the plan is to call a virtualenv python from filediver to create the blend file, similar to how ffmpeg is used.

DDS and EXR python scripts could be implemented in filediver directly, so that models are exported with EXR files included. However, the DDS files are just tiny LUTs, so conversion isn't the worst thing in the world. Some other programs may prefer to have them in the original DDS format anyway, and there isn't a ratified extension for EXR files in the GLTF standard