Include and use duplicate files
Filtering out duplicates has an adverse effect when extracting by triad: If the triad isn't the first place a file is defined, that file will not be exported with the rest of the triad. This is especially problematic when exporting entire armor sets, which can often share geometry. Now, exporting a triad should include every file present in the triad, even if it is duplicated elsewhere

Still need to figure out how material variants are applied, since many armors in game use materials that are different than the ones defined in their unit file