Correct fix for orphan meshes problem
Apparently skinned meshes were not having their mesh nodes added as children of the correct node, causing blender to count them as orphaned. The previous fix to add all mesh nodes to the scene was a workaround that caused issues when non-skinned meshes were added to the file, because their mesh nodes were correctly added as children.

Fixes #117