Update armor sets to use properly decrypted files rather than memory dumped files
When loaded into memory, datalibrary files have their pointers patched with the correct offset to be valid, dereferencable pointers. This was what necessitated the "armorSetOffset" to properly parse the files. I was unknowingly undoing the pointer patching the library was doing on load. Now that I can load these files from disk, there's no longer any patching going on and the offset is no longer needed.